fix: server Выделить логику инициализации хранилища в отдельный файл
